Security System-Sophomore Design
This is a self contained security system that was constructed as a college project, it includes a sensor to indicate a door is open and lights as indicators of the systems status. when the alarm is tripped a buzzer sounds and the red led flashes. A lcd screen displays commands and a keypad is used to input the password to arm and disarm the system.